Allow injection of SPIFFS read/write/erase functions at object instantiation of Arduino_SPIFFS object
