+++ src/Arduino_W25Q16DV.h
@@ -27,6 +27,11 @@
 
 #include <SPI.h>
 
+extern "C"
+{
+#include "spiffs.h"
+}
+
 /**************************************************************************************
  * TYPEDEF
  **************************************************************************************/
@@ -114,5 +119,13 @@ private:
  **************************************************************************************/
 
 extern Arduino_W25Q16DV flash;
+
+/**************************************************************************************
+ * FREE FUNCTION DECLARATION
+ **************************************************************************************/
+
+s32_t w25q16_spi_read (u32_t addr, u32_t size, u8_t * buf);
+s32_t w25q16_spi_write(u32_t addr, u32_t size, u8_t * buf);
+s32_t w25q16_spi_erase(u32_t addr, u32_t size);
 
 #endif /* ARDUINO_W25Q16DV_H_ */
